refactor: made `banded_linalg` private for now
