+++ chemotools/utils/_banded_linalg.py
